refactor: styled-component 제거 및 module sass로 변경
- remove emotion
- install sass dependencies
