+++ @types/declaration.d.ts
@@ -1,0 +1,4 @@
+declare module '*.scss' {
+  const content: Record<string, string>;
+  export default content;
+}
