+++ package.json
@@ -15,15 +15,14 @@
   "dependencies": {
     "@docusaurus/core": "^2.4.1",
     "@docusaurus/preset-classic": "^2.4.1",
-    "@emotion/core": "^11.0.0",
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
     "@mdx-js/react": "^1.6.22",
     "clsx": "^1.2.1",
+    "docusaurus-plugin-sass": "^0.2.5",
     "dotenv": "^16.0.3",
     "prism-react-renderer": "^1.3.5",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "sass": "^1.64.2",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
