+++ src/components/RowSummaryBooks/index.module.scss
@@ -1,0 +1,4 @@
+.rowSummaryBooksWrapper {
+  text-align: center;
+  align-items: flex-end;
+}
