+++ src/components/SummaryBookTemplate/index.module.scss
@@ -1,0 +1,3 @@
+.bookInfoWrapper {
+  margin-top: 1rem;
+}
