+++ src/components/SummaryBooks/index.module.scss
@@ -1,0 +1,6 @@
+.booksInfoSection {
+  display: flex;
+  align-items: center;
+  padding: 2rem 0 5rem 0;
+  width: 100%;
+}
