+++ src/components/SummaryBooksView/index.module.scss
@@ -1,0 +1,4 @@
+.mountainImg {
+  height: 200px;
+  width: 200px;
+}
