+++ src/components/SummaryBooksView/index.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 
-import styled from '@emotion/styled';
-
 import clsx from 'clsx';
 
 import { v4 as uuidv4 } from 'uuid';
 
-import books from '../data/books';
-import MountainSvg from '../../static/img/undraw_docusaurus_mountain.svg';
+import books from '@site/src/data/books';
+import MountainSvg from '../../../static/img/undraw_docusaurus_mountain.svg';
+import RowSummaryBooks from '../RowSummaryBooks';
 
-import RowSummaryBooksWrapper from './RowSummaryBooksWrapper';
+import styles from './index.module.scss';
 
 type Props = {
   title: string;
@@ -19,13 +18,13 @@ function SummaryBooksView({ title }: Props) {
   return (
     <div className={clsx('col')}>
       <div className="text--center">
-        <MountainImg />
+        <MountainSvg className={styles.mountainImg} />
       </div>
       <div className="text--center padding-horiz--md">
         <h2>{title}</h2>
         <div>
           {books.map((rowBooks) => (
-            <RowSummaryBooksWrapper
+            <RowSummaryBooks
               key={uuidv4()}
               rowBooks={rowBooks}
             />
@@ -37,8 +36,3 @@ function SummaryBooksView({ title }: Props) {
 }
 
 export default SummaryBooksView;
-
-const MountainImg = styled(MountainSvg)`
-  height: 200px;
-  width: 200px;
-`;
